Move the telemetry information to service
